Extract local transport
